Disable interactions with map from the contextual layer modal
