+++ src/components/map/controls/layer-toggle/index.js
@@ -1,3 +1,17 @@
+import React, { useState, useRef } from 'react';
 import Component from './component';
 
-export default Component;
+const LayerToggle = () => {
+  const [toggleLayerActive, setToggleLayerActive] = useState(false);
+  const tooltipRef = useRef(null);
+
+  return (
+    <Component
+      tooltipRef={tooltipRef}
+      toggleLayerActive={toggleLayerActive}
+      setToggleLayerActive={setToggleLayerActive}
+    />
+  );
+};
+
+export default LayerToggle;
