+++ src/components/map/controls/layer-toggle/styles.module.scss
@@ -1,0 +1,11 @@
+@import '~styles/settings';
+@import '../styles.scss';
+
+.navigationBar {
+  position: absolute;
+  left: 0;
+  top: 50%;
+  border: 1px solid #222222;
+  border-left: 0;
+  box-shadow: 0 1px 2px 0 rgba(0,0,0,0.09);
+}
