--- src/components/map/controls/layer-toggle/styles.scss
@@ -1,17 +1,0 @@
-.c-layer-toggle {
-  position: absolute;
-  bottom: 30px;
-  right: 10px;
-  background: #2b2d2f;
-  padding: 16px 32px;
-  border-radius: 4px;
-  font-family: monospace;
-
-  p {
-    color: #fafafa;
-
-    input {
-      margin-left: 8px;
-    }  
-  }
-}
