--- src/components/map/controls/navigation/styles.module.scss
@@ -1,42 +1,0 @@
-@import '~styles/settings';
-
-$navigation-button-size: 45px;
-
-.navigationButton {
-  background-color: white;
-  height: $navigation-button-size;
-  width: $navigation-button-size;
-
-  &:hover {
-    background-color: white;
-  }
-}
-
-.navigationBar {
-  position: absolute;
-  left: 0;
-  top: calc(50% - 1.5 * #{$navigation-button-size});
-  border: 1px solid #222222;
-  border-left: 0;
-  box-shadow: 0 1px 2px 0 rgba(0,0,0,0.09);
-
-  .zoomButtons {
-    border-radius: 0;
-    box-shadow: none !important;
-    
-    & > button {
-      @extend .navigationButton;
-      border-bottom: 1px solid #222222;
-      background-repeat: no-repeat;
-      background-position: center;
-
-      &:first-child {
-        background-image: url('../../../../assets/icons/zoomin.svg');
-      }
-
-      &:last-child {
-        background-image: url('../../../../assets/icons/zoomout.svg');    
-      }
-    }
-  }
-}
