+++ src/components/map/controls/zoom/styles.module.scss
@@ -1,0 +1,27 @@
+@import '~styles/settings';
+@import '../styles.scss';
+
+.zoomButtons {
+  border-radius: 0 !important;
+  box-shadow: none !important;
+  position: absolute;
+  top: calc(50% - 90px);
+  border: 1px solid #222222;
+  border-left: 0;
+  box-shadow: 0 1px 2px 0 rgba(0,0,0,0.09);
+
+  & > button {
+    @extend .navigationButton;
+    border-bottom: 1px solid #222222 !important;
+    background-repeat: no-repeat;
+    background-position: center;
+
+    &:first-child {
+      background-image: url('../../../../assets/icons/zoomin.svg');
+    }
+
+    &:last-child {
+      background-image: url('../../../../assets/icons/zoomout.svg');    
+    }
+  }
+}
